+++ src/main/java/baseball/Util/Game.java
@@ -1,0 +1,2 @@
+package baseball.Util;public class Game {
+}
