+++ src/main/java/baseball/view/GameDisplay.java
@@ -1,0 +1,2 @@
+package baseball.view;public class GameDisplay {
+}
